% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bayes.sample}
\alias{bayes.sample}
\title{Example output of dada2 assignTaxonomy function}
\format{
^ list with 2 elements
\describe{
\item{tax}{taxonomic assignments}
\item{boot}{bootstrap confidence estimates}
}
}
\usage{
bayes.sample
}
\description{
Example output of dada2 assignTaxonomy function
}
\keyword{datasets}
