% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gg_13_8_train_set_97}
\alias{gg_13_8_train_set_97}
\title{All unique taxonomic assignments from the GreenGenes v13.8 clusted at 97\%}
\format{
^ dataframe with 4163 rows and 7 columns
\describe{
\item{domain}{domain assignments}
\item{phylum}{phylum assignments}
\item{class}{class assignments}
\item{order}{order assignments}
\item{family}{family assignments}
\item{genus}{genus assignments}
\item{species}{genus assignments}
}
}
\usage{
gg_13_8_train_set_97
}
\description{
All unique taxonomic assignments from the GreenGenes v13.8 clusted at 97\%
}
\keyword{datasets}
