% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{idtax.pr2.sample}
\alias{idtax.pr2.sample}
\title{Example output of DECIPHER idtaxa function with pr2 taxonomy}
\format{
^ list with 5 elements
\describe{
\item{1}{tax data for ASV 1}
\item{2}{tax data for ASV 2}
\item{3}{tax data for ASV 3}
\item{4}{tax data for ASV 4}
\item{5}{tax data for ASV 5}
}
}
\usage{
idtax.pr2.sample
}
\description{
Example output of DECIPHER idtaxa function with pr2 taxonomy
}
\keyword{datasets}
