% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pr2v4.12.0}
\alias{pr2v4.12.0}
\title{All unique taxonomic assignments from the pr2 reference database v4.12.0}
\format{
^ dataframe with 45352 rows and 8 columns
\describe{
\item{kingdom}{kingdom assignments}
\item{supergroup}{supergroup assignments}
\item{division}{division assignments}
\item{class}{class assignments}
\item{order}{order assignments}
\item{family}{family assignments}
\item{genus}{genus assignments}
\item{species}{species assignments}
}
}
\usage{
pr2v4.12.0
}
\description{
All unique taxonomic assignments from the pr2 reference database v4.12.0
}
\keyword{datasets}
