% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdp_train_set_16}
\alias{rdp_train_set_16}
\title{All unique taxonomic assignments from the RDP Train Set 16}
\format{
^ dataframe with 2472 rows and 6 columns
\describe{
\item{domain}{domain assignments}
\item{phylum}{phylum assignments}
\item{class}{class assignments}
\item{order}{order assignments}
\item{family}{family assignments}
\item{genus}{genus assignments}
}
}
\usage{
rdp_train_set_16
}
\description{
All unique taxonomic assignments from the RDP Train Set 16
}
\keyword{datasets}
