% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rubric.sample}
\alias{rubric.sample}
\title{Example rubric with ASV-identifying data}
\format{
^ DNAStringSet with 5 elements
\describe{
\item{sv1}{sample ASV 1}
\item{sv2}{sample ASV 2}
\item{sv3}{sample ASV 3}
\item{sv4}{sample ASV 4}
\item{sv5}{sample ASV 5}
}
}
\usage{
rubric.sample
}
\description{
Example rubric with ASV-identifying data
}
\keyword{datasets}
