% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{silva.nr.v138}
\alias{silva.nr.v138}
\title{All unique taxonomic assignments from the Silva SSU nr database v138}
\format{
^ dataframe with 6011 rows and 6 columns
\describe{
\item{domain}{domain assignments}
\item{phylum}{phylum assignments}
\item{class}{class assignments}
\item{order}{order assignments}
\item{family}{family assignments}
\item{genus}{genus assignments}
}
}
\usage{
silva.nr.v138
}
\description{
All unique taxonomic assignments from the Silva SSU nr database v138
}
\keyword{datasets}
