% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synonyms_v2}
\alias{synonyms_v2}
\title{Taxonomic synonyms searched by the taxmapper algorithm}
\format{
^ dataframe with 174 rows and 11 columns
\describe{
\item{Name_1}{first synonym}
\item{Name_2}{second synonym}
\item{Name_3}{third synonym}
\item{Name_4}{fourth synonym}
\item{Name_5}{fifth synonym}
\item{Name_6}{sixth synonym}
\item{Name_7}{seventh synonym}
\item{References}{Reference for some synonyms}
\item{Notes.References}{Notes from references}
\item{X}{Additional references for some synonyms}
\item{X.1}{Additional references for some synonyms}
}
}
\usage{
synonyms_v2
}
\description{
Taxonomic synonyms searched by the taxmapper algorithm
}
\keyword{datasets}
