% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmapper.R
\name{taxmapper}
\alias{taxmapper}
\title{Maps an input taxonomy table onto a different taxonomic nomenclature.}
\usage{
taxmapper(
  tt,
  tt.ranks = colnames(tt),
  tax2map2 = "pr2",
  exceptions = c("Archaea", "Bacteria"),
  ignore.format = FALSE,
  synonym.file = "default",
  streamline = TRUE,
  outfilez = NULL
)
}
\arguments{
\item{tt}{The input taxonomy table you would like to map onto a new
taxonomic nomenclature. Should be a dataframe of type char (no factors).}

\item{tt.ranks}{A character vector of the column names where taxonomic
names are found in tt. Supply them heirarchically (e.g. kingdom --> species)}

\item{tax2map2}{The taxonomic nomenclature you would like to map onto. pr2
v4.12.0 and Silva SSU v138 nr taxonomic nomenclatures are included in the
ensembleTax package. You can map to these by specifying "pr2" or "Silva".
Otherwise should be a dataframe of type character (no factors) with each
column corresponding to a taxonomic rank.}

\item{exceptions}{A character vector of taxonomic names at the basal/root
rank of tt that will be propagated onto the mapped taxonomy. ASVs assigned
to these names will retain these names at their basal/root rank in the mapped
taxonomy. All other ranks are assigned NA.}

\item{ignore.format}{If TRUE, the algorithm modifies taxonomic names to
account for common variations in taxonomic name syntax and/or formatting
commonly encountered in reference databases (e.g. Pseudonitzschia will map to
Pseudo-nitzschia). If FALSE, formatting issues may preclude mapping of
synonymous taxonomic names (e.g. Pseudonitzschia will NOT map to
Pseudo-nitzschia). An exhaustive list of formatting details is included in
Details.}

\item{synonym.file}{If "default", taxmapper uses taxonomic synonyms included
with the ensembleTax package. If a custom taxonomic synonym file is
preferred, a string corresponding to the name of the csv file should be
supplied. Taxonomic synonyms are searched when exact name matches are not
found in tax2map2. ignore.format does not apply to synonyms.}

\item{streamline}{If TRUE, only the mapped version of tt is returned as a
dataframe. If FALSE, a 3-element list is returned where element 1 is the
mapping rubric returned as a dataframe, element 2 is a character vector of
all names that could not be mapped (no exact matches found in tax2map2), and
element 3 is the mapped version of tt (a dataframe).}

\item{outfilez}{If NULL, mapping files are not saved to the current working
directory. Otherwise should be a 3-element character vector including, in
this order, the name of the file to store the taxonomic mapping key, the name
of the file to store the names that could not be mapped, and the name of the
file to store the ASVs supplied with tt with their mapped taxonomic
assignments. Each element of the vector should end in csv (only csv files
may be saved)}
}
\description{
Maps an input taxonomy table onto a different taxonomic nomenclature.
}
\details{
Exceptions should be used when the user knows a particular taxonomic
group is not found in tax2map2. The user is responsible for supplying valid
taxonomic names as these must be found in tt and will be propagated as
given to all ASVs that are assigned this name in tt. This should only be
used for high-level taxonomic groups that are not found in a database (e.g.
for retaining Eukaryota when mapping onto a prokaryote-only taxonomic
nomenclature).

When ignore.format = TRUE, taxmapper removes hyphens, underscores, and spaces
from all elements of tt. Each unique element separated by a hyphen,
underscore, or space, in addition to a concatenated string with each
separate sub-string combined into one, are compiled. It then creates
all-lower and all-upper case versions of these elements, and adds them to the
collection of unique elements. Each of these elements is then searched for
exact name matches in tax2map2. The following taxonomic suffixes are
equivalent when ignore.format = TRUE: phyta, phytes, phyte, and phyceae.
ignore.format is never applied to synonym.file.

For high-throughput implementation of taxmapper, it's recommended to set
streamline = TRUE.
}
\examples{
fake.silva <- data.frame(ASV = c("AAAA", "ATCG", "GCGC", "TATA", "TCGA"),
domain = c("Bacteria", "Eukaryota", "Eukaryota", "Eukaryota", "Eukaryota"),
phylum = c("Firmicutes", "Diatomea", "Retaria", "MAST-12", "Diatomea"),
class = c(NA, "Coscinodiscophytina_cl", "Polycystinea", "MAST-12A",
"Mediophyceae"),
order = c(NA, "Fragilariales", "Collodaria", NA, NA),
family = c(NA, "Fragilariales_fa", "Collodaria_fa", NA, NA),
genus = c(NA, "Podocystis", "Collophidium", NA, NA),
stringsAsFactors = FALSE)
head(fake.silva)
mapped.silva <- taxmapper(fake.silva,
                          tt.ranks = colnames(fake.silva)[2:ncol(fake.silva)],
                          tax2map2 = "pr2",
                          exceptions = c("Archaea", "Bacteria"),
                          ignore.format = FALSE,
                          synonym.file = "default",
                          streamline = TRUE,
                          outfilez = NULL)

}
\seealso{
idtax2df, bayestax2df, ensembleTax
}
\author{
Dylan Catlett

Kevin Son
}
