% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UNITE.euk.10.05.2021}
\alias{UNITE.euk.10.05.2021}
\title{All unique taxonomic assignments from the UNITE all eukaryote database}
\format{
^ dataframe with 66383 rows and 7 columns
\describe{
\item{kingdom}{kingdom assignments}
\item{phylum}{phylum assignments}
\item{class}{class assignments}
\item{order}{order assignments}
\item{family}{family assignments}
\item{genus}{genus assignments}
\item{species}{species assignments}
}
}
\usage{
UNITE.euk.10.05.2021
}
\description{
All unique taxonomic assignments from the UNITE all eukaryote database
}
\keyword{datasets}
